+++ regalia/include/CreditState.h
@@ -11,7 +11,7 @@ public:
 
 	void LoadAssets() override;
 
-	void Update(float dt) override;
+	void Update(unsigned dt) override;
 
 	void Render() override;
