+++ regalia/include/PlayerAim.h
@@ -7,7 +7,7 @@ public:
 	
 	~PlayerAim();
 
-	void Update(float dt) override;
+	void Update(unsigned dt) override;
 	
 	void Render() override;
