+++ regalia/include/SpriteSheetData.h
@@ -7,11 +7,11 @@ public:
 
 	std::string file;
 	
-	float frameTime;
+	unsigned frameTime;
 	
 	bool selfDestruct;
 	
-	SpriteSheetData(const std::string& file, int imageWidth, int imageHeight, float frameTime, int totalAnimations, const std::map<int, int>& animationsFrames, bool selfDestruct = false);
+	SpriteSheetData(const std::string& file, int imageWidth, int imageHeight, unsigned frameTime, int totalAnimations, const std::map<int, int>& animationsFrames, bool selfDestruct = false);
 	
 	SpriteSheetData(const SpriteSheetData& other) = delete;
