+++ regalia/src/SpriteSheetData.cpp
@@ -2,7 +2,7 @@
 #include <SpriteSheetData.h>
 #include <Logger.h>
 
-SpriteSheetData::SpriteSheetData(const std::string& file, int imageWidth, int imageHeight, float frameTime, int totalAnimations, const std::map<int, int>& animationsFrames, bool selfDestruct)
+SpriteSheetData::SpriteSheetData(const std::string& file, int imageWidth, int imageHeight, unsigned frameTime, int totalAnimations, const std::map<int, int>& animationsFrames, bool selfDestruct)
     : file(file)
     , frameTime(frameTime)
     , selfDestruct(selfDestruct)
